requirements.txt
alembic==1.14.0
asgiref==3.8.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.0
coverage==7.6.9
crispy-bootstrap5==0.7
cycler==0.12.1
databricks-sdk==0.38.0
Deprecated==1.2.15
Django==4.2.7
django-crispy-forms==2.1
docker==7.1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==33.1.0
Flask==3.1.0
fonttools==4.55.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mlflow==2.18.0
mlflow-skinny==2.18.0
numpy==1.26.4
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pandas==2.1.3
pillow==11.0.0
plotly==5.18.0
pluggy==1.5.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.13.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
tenacity==9.0.0
threadpoolctl==3.5.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
whitenoise==6.6.0
wrapt==1.17.0
zipp==3.21.0
seaborn==0.12.2
psutil== 7.0.0
prometheus_client==0.21.1
drf-spectacular==0.27.0
drf-spectacular-sidecar
django-prometheus
psycopg2-binary==2.9.7
djangorestframework-simplejwt>=5.3.0
